--- app/src/main/java/com/sw/wordgarden/presentation/ui/quiz/QuizFragment.kt
@@ -1,20 +1,0 @@
-package com.sw.wordgarden.presentation.ui.quiz
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.sw.wordgarden.R
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class QuizFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_quiz, container, false)
-    }
-}
