+++ app/src/main/java/com/sw/wordgarden/presentation/ui/quiz/makequiz/MakeQuizFragment.kt
@@ -1,0 +1,29 @@
+package com.sw.wordgarden.presentation.ui.quiz.makequiz
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.sw.wordgarden.databinding.FragmentMakeQuizBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MakeQuizFragment : Fragment() {
+
+    private var _binding: FragmentMakeQuizBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentMakeQuizBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
