+++ app/src/main/java/com/sw/wordgarden/presentation/ui/quiz/quiz/QuizFragment.kt
@@ -1,0 +1,29 @@
+package com.sw.wordgarden.presentation.ui.quiz.quiz
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.sw.wordgarden.databinding.FragmentQuizBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class QuizFragment : Fragment() {
+
+    private var _binding: FragmentQuizBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentQuizBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
