+++ app/src/main/res/layout/fragment_make_quiz.xml
@@ -2,10 +2,11 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/cl_make_quiz"
+    android:id="@+id/cl_make_quiz_inner"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/light_khaki">
+    android:background="@color/light_khaki"
+    tools:context=".presentation.ui.quiz.makequiz.MakeQuizFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_make_quiz_top_bar"
